Use updated plugin
