+++ pxtblocks/index.ts
@@ -19,6 +19,7 @@ export * from "./importer";
 export * from "./diff";
 export * from "./legacyMutations";
 export * from "./blockDragger";
+export * from "./workspaceSearch";
 
 import * as contextMenu from "./contextMenu";
 import * as external from "./external";
